Terminal Progress bar with image Resizing/requirements.txt
Pillow
tqdm==4.48.2